ATR-753: reworked the code fix into three separate code actions
